@@ -1,5 +1,5 @@
-name: card_scanner_example
-description: Demonstrates how to use the card_scanner plugin.
+name: credit_card_scanner_example
+description: Demonstrates how to use the credit_card_scanner plugin.
 version: 0.2.2+5
 
 # The following line prevents the package from being accidentally published to
@@ -7,15 +7,15 @@ version: 0.2.2+5
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  card_scanner:
+  credit_card_scanner:
     # When depending on this package from a real application you should use:
-    #   card_scanner: ^x.y.z
+    #   credit_card_scanner: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version. 
@@ -27,6 +27,7 @@ dependencies:
 
 dev_dependencies:
   flutter_launcher_icons: "^0.7.3"
+  flutter_lints: ^2.0.0
 
   flutter_test:
     sdk: flutter
