@@ -1,26 +1,29 @@
-name: card_scanner
-description: Fast, Accurate and Secure Credit & Debit card scanning library for flutter.
-version: 1.0.1-prerelease
+name: credit_card_scanner
+description: Fast, Accurate and Secure Credit & Debit card scanning library for flutter. This package is a fork of the original.
+version: 1.0.5
 homepage: https://github.com/nateshmbhat/card-scanner-flutter
 repository: https://github.com/nateshmbhat/card-scanner-flutter
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  
+  flutter_lints: ^2.0.0
 
 flutter:
- plugin:
+  plugin:
     platforms:
       android:
-        package: com.nateshmbhat.card_scanner
+        package: com.nateshmbhat.credit_card_scanner
         pluginClass: CardScannerPlugin
       ios:
         pluginClass: CardScannerPlugin
