@@ -1,23 +1,21 @@
-import 'dart:math';
-
-import 'package:card_scanner_example/scan_option_configure_widget/scan_option_configure_widget.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
 
-import 'package:card_scanner/card_scanner.dart';
+import 'package:credit_card_scanner/credit_card_scanner.dart';
+import 'scan_option_configure_widget/scan_option_configure_widget.dart';
 
-void main() {
-  runApp(MyApp());
-}
+void main() => runApp( const MyApp() );
 
 class MyApp extends StatefulWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
-  CardDetails _cardDetails;
-  CardScanOptions scanOptions = CardScanOptions(
+  CardDetails? _cardDetails;
+  CardScanOptions scanOptions = const CardScanOptions(
     scanCardHolderName: true,
     // enableDebugLogs: true,
     validCardsToScanBeforeFinishingScan: 5,
@@ -27,8 +25,8 @@ class _MyAppState extends State<MyApp> {
   );
 
   Future<void> scanCard() async {
-    var cardDetails = await CardScanner.scanCard(scanOptions: scanOptions);
-    if (!mounted) return;
+    final CardDetails? cardDetails = await CardScanner.scanCard(scanOptions: scanOptions);
+    if ( !mounted || cardDetails == null ) return;
     setState(() {
       _cardDetails = cardDetails;
     });
@@ -39,17 +37,18 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('card_scanner app'),
+          title: const Text('credit_card_scanner app'),
         ),
         body: Center(
           child: Column(
             mainAxisAlignment: MainAxisAlignment.spaceEvenly,
             children: [
-              RaisedButton(
+              MaterialButton(
+                color: Colors.blue,
                 onPressed: () async {
                   scanCard();
                 },
-                child: Text('scan card'),
+                child: const Text('scan card'),
               ),
               Text('$_cardDetails'),
               Expanded(
