@@ -1,26 +1,28 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'dart:async';
 
-import 'package:card_scanner/card_scanner.dart';
+import 'package:credit_card_scanner/credit_card_scanner.dart';
 
 class OptionConfigureWidget extends StatefulWidget {
-  final void Function(CardScanOptions scanOptions) onScanOptionChanged;
-  final CardScanOptions initialOptions;
+  final void Function(CardScanOptions scanOptions)? onScanOptionChanged;
+  final CardScanOptions? initialOptions;
 
-  const OptionConfigureWidget({Key key, @required this.onScanOptionChanged, this.initialOptions}) : super(key: key);
+  const OptionConfigureWidget({Key? key, @required this.onScanOptionChanged, this.initialOptions}) : super(key: key);
 
   @override
-  _OptionConfigureWidgetState createState() => _OptionConfigureWidgetState();
+  State<OptionConfigureWidget> createState() => _OptionConfigureWidgetState();
 }
 
 class _OptionConfigureWidgetState extends State<OptionConfigureWidget> {
   bool scanExpiryDate = true;
+
   bool scanCardHolderName = true;
 
   int initialScansToDrop = 1;
+
   int validCardsToScanBeforeFinishingScan = 6;
+
   List<String> cardHolderNameBlackListedWords = [];
 
   bool considerPastDatesInExpiryDateScan = false;
@@ -39,7 +41,7 @@ class _OptionConfigureWidgetState extends State<OptionConfigureWidget> {
   void initState() {
     if (widget.initialOptions != null) {
       final options = widget.initialOptions;
-      scanExpiryDate = options.scanExpiryDate;
+      scanExpiryDate = options!.scanExpiryDate;
       scanCardHolderName = options.scanCardHolderName;
       initialScansToDrop = options.initialScansToDrop;
       validCardsToScanBeforeFinishingScan = options.validCardsToScanBeforeFinishingScan;
@@ -75,26 +77,28 @@ class _OptionConfigureWidgetState extends State<OptionConfigureWidget> {
               possibleCardHolderNamePositions.contains(CardHolderNameScanPosition.aboveCardNumber), (newValue) {
             if (newValue == true) {
               possibleCardHolderNamePositions.add(CardHolderNameScanPosition.aboveCardNumber);
-            } else
+            } else {
               possibleCardHolderNamePositions.remove(CardHolderNameScanPosition.aboveCardNumber);
+            }
             setState(() {});
           }),
           buildCheckBox('expect card holder BELOW card number',
               possibleCardHolderNamePositions.contains(CardHolderNameScanPosition.belowCardNumber), (newValue) {
             if (newValue == true) {
               possibleCardHolderNamePositions.add(CardHolderNameScanPosition.belowCardNumber);
-            } else
+            } else {
               possibleCardHolderNamePositions.remove(CardHolderNameScanPosition.belowCardNumber);
+            }
             setState(() {});
           }),
-          Divider(),
+          const Divider(),
           Container(
-            padding: EdgeInsets.all(20),
+            padding: const EdgeInsets.all(20),
             child: Column(
               children: [
-                Text('black listed card holder names (comma separated)'),
+                const Text('black listed card holder names (comma separated)'),
                 TextField(
-                  decoration: InputDecoration(border: OutlineInputBorder()),
+                  decoration: const InputDecoration(border: OutlineInputBorder()),
                   onChanged: (value) => cardHolderNameBlackListedWords = value.split(','),
                   onEditingComplete: () {
                     setState(() {});
@@ -111,7 +115,7 @@ class _OptionConfigureWidgetState extends State<OptionConfigureWidget> {
   Widget buildCheckBox(String key, bool value, onChanged) {
     return Column(
       children: [
-        Divider(),
+        const Divider(),
         Row(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
@@ -131,21 +135,21 @@ class _OptionConfigureWidgetState extends State<OptionConfigureWidget> {
   Widget buildIntegerEditWidget(String key, int value, onChanged) {
     return Column(
       children: [
-        Divider(),
+        const Divider(),
         Row(
           mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
             Text(key),
             Text(value.toString()),
             IconButton(
-              icon: Icon(Icons.arrow_upward),
+              icon: const Icon(Icons.arrow_upward),
               onPressed: () {
                 onChanged(value + 1);
                 setState(() {});
               },
             ),
             IconButton(
-              icon: Icon(Icons.arrow_downward),
+              icon: const Icon(Icons.arrow_downward),
               onPressed: () {
                 onChanged(max(0, value - 1));
                 setState(() {});
@@ -159,7 +163,7 @@ class _OptionConfigureWidgetState extends State<OptionConfigureWidget> {
 
   @override
   void setState(fn) {
-    widget.onScanOptionChanged(CardScanOptions(
+    widget.onScanOptionChanged!(CardScanOptions(
         possibleCardHolderNamePositions: possibleCardHolderNamePositions.toList(),
         cardHolderNameBlackListedWords: cardHolderNameBlackListedWords,
         enableDebugLogs: enableDebugLogs,
